refactored graph conans->conan
