+++ conan/api/subapi/lockfile.py
@@ -2,7 +2,7 @@ import os
 
 from conan.api.output import ConanOutput
 from conan.cli import make_abs_path
-from conans.client.graph.graph import Overrides
+from conan.internal.graph.graph import Overrides
 from conan.errors import ConanException
 from conan.internal.model.lockfile import Lockfile, LOCKFILE
