+++ conan/internal/graph/compatibility.py
@@ -3,7 +3,7 @@ from collections import OrderedDict
 
 from conan.api.output import ConanOutput
 from conan.internal.cache.home_paths import HomePaths
-from conans.client.graph.compute_pid import run_validate_package_id
+from conan.internal.graph.compute_pid import run_validate_package_id
 from conans.client.loader import load_python_file
 from conan.internal.errors import conanfile_exception_formatter, scoped_traceback
 from conan.errors import ConanException
