+++ conan/internal/graph/graph_builder.py
@@ -5,13 +5,13 @@ from collections import deque
 from conan.internal.cache.conan_reference_layout import BasicLayout
 from conan.internal.methods import run_configure_method
 from conan.internal.model.recipe_ref import ref_matches
-from conans.client.graph.graph import DepsGraph, Node, CONTEXT_HOST, \
+from conan.internal.graph.graph import DepsGraph, Node, CONTEXT_HOST, \
     CONTEXT_BUILD, TransitiveRequirement, RECIPE_VIRTUAL, RECIPE_EDITABLE
-from conans.client.graph.graph import RECIPE_PLATFORM
-from conans.client.graph.graph_error import GraphLoopError, GraphConflictError, GraphMissingError, \
+from conan.internal.graph.graph import RECIPE_PLATFORM
+from conan.internal.graph.graph_error import GraphLoopError, GraphConflictError, GraphMissingError, \
     GraphRuntimeError, GraphError
-from conans.client.graph.profile_node_definer import initialize_conanfile_profile
-from conans.client.graph.provides import check_graph_provides
+from conan.internal.graph.profile_node_definer import initialize_conanfile_profile
+from conan.internal.graph.provides import check_graph_provides
 from conan.errors import ConanException
 from conan.internal.model.conan_file import ConanFile
 from conan.internal.model.options import Options, _PackageOptions
