+++ conan/internal/graph/proxy.py
@@ -1,8 +1,8 @@
 from conan.api.output import ConanOutput
 from conan.internal.cache.conan_reference_layout import BasicLayout
-from conans.client.graph.graph import (RECIPE_DOWNLOADED, RECIPE_INCACHE, RECIPE_NEWER,
-                                       RECIPE_NOT_IN_REMOTE, RECIPE_UPDATED, RECIPE_EDITABLE,
-                                       RECIPE_INCACHE_DATE_UPDATED, RECIPE_UPDATEABLE)
+from conan.internal.graph.graph import (RECIPE_DOWNLOADED, RECIPE_INCACHE, RECIPE_NEWER,
+                                        RECIPE_NOT_IN_REMOTE, RECIPE_UPDATED, RECIPE_EDITABLE,
+                                        RECIPE_INCACHE_DATE_UPDATED, RECIPE_UPDATEABLE)
 from conan.internal.errors import NotFoundException
 from conan.errors import ConanException
