+++ conan/internal/model/dependencies.py
@@ -1,6 +1,6 @@
 from collections import OrderedDict
 
-from conans.client.graph.graph import RECIPE_PLATFORM
+from conan.internal.graph.graph import RECIPE_PLATFORM
 from conan.errors import ConanException
 from conan.api.model import RecipeReference
 from conan.internal.model.conanfile_interface import ConanFileInterface
