+++ conan/tools/cmake/presets.py
@@ -9,7 +9,7 @@ from conan.tools.cmake.toolchain.blocks import GenericSystemBlock
 from conan.tools.cmake.utils import is_multi_configuration
 from conan.tools.build import build_jobs
 from conan.tools.microsoft import is_msvc
-from conans.client.graph.graph import RECIPE_CONSUMER
+from conan.internal.graph.graph import RECIPE_CONSUMER
 from conan.errors import ConanException
 from conans.util.files import save, load
