+++ conan/tools/cmake/toolchain/toolchain.py
@@ -22,7 +22,7 @@ from conan.tools.microsoft import VCVars
 from conan.tools.microsoft.visual import vs_ide_version
 from conan.errors import ConanException
 from conan.internal.model.options import _PackageOption
-from conans.client.graph.graph import RECIPE_CONSUMER, RECIPE_EDITABLE
+from conan.internal.graph.graph import RECIPE_CONSUMER, RECIPE_EDITABLE
 from conans.util.files import save
