+++ test/integration/graph/core/test_build_requires.py
@@ -4,7 +4,7 @@ import pytest
 
 from parameterized import parameterized
 
-from conans.client.graph.graph_error import GraphConflictError, GraphLoopError, GraphRuntimeError
+from conan.internal.graph.graph_error import GraphConflictError, GraphLoopError, GraphRuntimeError
 from conan.api.model import RecipeReference
 from test.integration.graph.core.graph_manager_base import GraphManagerTest
 from conan.test.utils.tools import GenConanfile, NO_SETTINGS_PACKAGE_ID, TestClient
