+++ test/integration/graph/core/test_provides.py
@@ -2,7 +2,7 @@ import textwrap
 
 from parameterized import parameterized
 
-from conans.client.graph.graph_error import GraphProvidesError
+from conan.internal.graph.graph_error import GraphProvidesError
 from test.integration.graph.core.graph_manager_base import GraphManagerTest
 from test.integration.graph.core.graph_manager_test import _check_transitive
 from conan.test.utils.tools import GenConanfile, TestClient
