+++ test/integration/package_id/test_validate.py
@@ -8,7 +8,7 @@ import unittest
 import pytest
 
 from conan.cli.exit_codes import ERROR_INVALID_CONFIGURATION, ERROR_GENERAL
-from conans.client.graph.graph import BINARY_INVALID
+from conan.internal.graph.graph import BINARY_INVALID
 from conan.test.assets.genconanfile import GenConanfile
 from conans.util.files import save, load
 from conan.test.utils.tools import TestClient, NO_SETTINGS_PACKAGE_ID
